pyproject.toml
[tool.poetry]
name = "payhere-backend"
version = "0.1.0"
description = ""
authors = ["HoJin9622 <kiss0104040@gmail.com>"]
readme = "README.md"
packages = [{include = "payhere_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-environ = "^0.9.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
